requirements.txt
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudevents==1.6.1
dacite==1.6.0
Deprecated==1.2.13
deprecation==2.1.0
Faker==15.0.0
Flask==2.2.2
functions-framework==3.2.0
google-api-core==2.10.1
google-auth==2.12.0
google-cloud-bigquery==3.3.3
google-cloud-bigquery-storage==2.16.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
idna==3.4
inflect==6.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonlines==3.1.0
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
pluggy==1.0.0
proto-plus==1.22.1
protobuf==4.21.7
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.55
PyJWT==2.5.0
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
rsa==4.9
six==1.16.0
stringcase==1.2.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
watchdog==2.1.9
Werkzeug==2.2.2
wrapt==1.14.1
